ai-agent/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
redis[asyncio]>=5.0.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pydantic>=2.6.0
tavily-python>=0.3.0
langsmith>=0.1.0
